app/backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary>==2.9.1
flask-cors==4.0.1
gunicorn==20.1.0